got sd card working, trying to read audio files might have truble with ram if we read too much tho

+++ VisiTune/Core/Inc/sd_card.h
@@ -11,7 +11,7 @@
 #include "main.h"
 #include "ff.h"
 
-void read_sound_file(const TCHAR* file_name, uint8_t *pData, uint32_t dataSize);
+void read_sound_file(const TCHAR* file_name, uint16_t *pData, uint32_t dataSize);
